Add files via upload
OK, I had to delete the branch from before. because I am completely clueless how to push code from VSCode.  

Ended creating this branch (similar to before)
with fresh code that contains the images for each container of the clone static page. 

As we move forward I hope we can keep uploading changes here and ultimately push all of the code to the main branch. 
I think that is how is done (?) haha..